+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add validation for `cilium.giantswarm.io/ipam-mode` annotation on `Cluster` CR creation.
+- Default `cilium.giantswarm.io/ipam-mode` annotation on `Cluster` CR during creation of v19 cluster.
 
 ## [4.8.1] - 2023-06-07